- Updated emulator
git-svn-id: svn://localhost/mosa@1453 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/EmulatedKernel/MemoryDispatch.cs
@@ -22,11 +22,11 @@ namespace Mosa.EmulatedKernel
 		/// <summary>
 		/// 
 		/// </summary>
-		public static byte CR0;
+		public static uint CR0;
 		/// <summary>
 		/// 
 		/// </summary>
-		public static byte CR3;
+		public static uint CR3;
 
 		/// <summary>
 		///
